move lse solvers to analysis, to avoid circular import (if they use cpp functions
